+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - AI Guard: config APIs.
+- Prompt Guard: config APIs.
 
 ## 6.1.1 - 2025-05-12